requirements.txt
asgiref==3.9.1
boto3==1.40.34
botocore==1.40.34
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
jmespath==1.0.1
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
